requirements.txt

python-telegram-bot==13.15
feedparser


